+++ spec/api-net-spec.js
@@ -66,217 +66,6 @@ describe('net module', () => {
       session.defaultSession.webRequest.onBeforeRequest(null)
     })
 
-    it('should be able to remove a custom HTTP request header before first write', (done) => {
-      const requestUrl = '/requestUrl'
-      const customHeaderName = 'Some-Custom-Header-Name'
-      const customHeaderValue = 'Some-Customer-Header-Value'
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case requestUrl:
-            expect(request.headers[customHeaderName.toLowerCase()]).to.be.undefined()
-            response.statusCode = 200
-            response.statusMessage = 'OK'
-            response.end()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-      const urlRequest = net.request({
-        method: 'GET',
-        url: `${server.url}${requestUrl}`
-      })
-      urlRequest.on('response', (response) => {
-        const statusCode = response.statusCode
-        expect(statusCode).to.equal(200)
-        response.pause()
-        response.on('data', (chunk) => {
-        })
-        response.on('end', () => {
-          done()
-        })
-        response.resume()
-      })
-      urlRequest.setHeader(customHeaderName, customHeaderValue)
-      expect(urlRequest.getHeader(customHeaderName)).to.equal(customHeaderValue)
-      urlRequest.removeHeader(customHeaderName)
-      expect(urlRequest.getHeader(customHeaderName)).to.be.undefined()
-      urlRequest.write('')
-      urlRequest.end()
-    })
-
-    it('should not be able to remove a custom HTTP request header after first write', (done) => {
-      const requestUrl = '/requestUrl'
-      const customHeaderName = 'Some-Custom-Header-Name'
-      const customHeaderValue = 'Some-Customer-Header-Value'
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case requestUrl:
-            expect(request.headers[customHeaderName.toLowerCase()]).to.equal(customHeaderValue)
-            response.statusCode = 200
-            response.statusMessage = 'OK'
-            response.end()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-      const urlRequest = net.request({
-        method: 'GET',
-        url: `${server.url}${requestUrl}`
-      })
-      urlRequest.on('response', (response) => {
-        const statusCode = response.statusCode
-        expect(statusCode).to.equal(200)
-        response.pause()
-        response.on('data', (chunk) => {
-        })
-        response.on('end', () => {
-          done()
-        })
-        response.resume()
-      })
-      urlRequest.setHeader(customHeaderName, customHeaderValue)
-      expect(urlRequest.getHeader(customHeaderName)).to.equal(customHeaderValue)
-      urlRequest.write('')
-      expect(() => {
-        urlRequest.removeHeader(customHeaderName)
-      }).to.throw()
-      expect(urlRequest.getHeader(customHeaderName)).to.equal(customHeaderValue)
-      urlRequest.end()
-    })
-
-    it('should be able to set cookie header line', (done) => {
-      const requestUrl = '/requestUrl'
-      const cookieHeaderName = 'Cookie'
-      const cookieHeaderValue = 'test=12345'
-      const customSession = session.fromPartition('test-cookie-header')
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case requestUrl:
-            expect(request.headers[cookieHeaderName.toLowerCase()]).to.equal(cookieHeaderValue)
-            response.statusCode = 200
-            response.statusMessage = 'OK'
-            response.end()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-
-      customSession.cookies.set({
-        url: `${server.url}`,
-        name: 'test',
-        value: '11111'
-      }).then(() => { // resolved
-        const urlRequest = net.request({
-          method: 'GET',
-          url: `${server.url}${requestUrl}`,
-          session: customSession
-        })
-        urlRequest.on('response', (response) => {
-          const statusCode = response.statusCode
-          expect(statusCode).to.equal(200)
-          response.pause()
-          response.on('data', (chunk) => {})
-          response.on('end', () => {
-            done()
-          })
-          response.resume()
-        })
-        urlRequest.setHeader(cookieHeaderName, cookieHeaderValue)
-        expect(urlRequest.getHeader(cookieHeaderName)).to.equal(cookieHeaderValue)
-        urlRequest.end()
-      }, (error) => {
-        done(error)
-      })
-    })
-
-    it('should be able to abort an HTTP request before first write', (done) => {
-      const requestUrl = '/requestUrl'
-      server.on('request', (request, response) => {
-        response.end()
-        expect.fail('Unexpected request event')
-      })
-
-      let requestAbortEventEmitted = false
-      let requestCloseEventEmitted = false
-
-      const urlRequest = net.request({
-        method: 'GET',
-        url: `${server.url}${requestUrl}`
-      })
-      urlRequest.on('response', (response) => {
-        expect.fail('Unexpected response event')
-      })
-      urlRequest.on('finish', () => {
-        expect.fail('Unexpected finish event')
-      })
-      urlRequest.on('error', () => {
-        expect.fail('Unexpected error event')
-      })
-      urlRequest.on('abort', () => {
-        requestAbortEventEmitted = true
-      })
-      urlRequest.on('close', () => {
-        requestCloseEventEmitted = true
-        expect(requestAbortEventEmitted).to.be.true()
-        expect(requestCloseEventEmitted).to.be.true()
-        done()
-      })
-      urlRequest.abort()
-      expect(urlRequest.write('')).to.be.false()
-      urlRequest.end()
-    })
-
-    it('it should be able to abort an HTTP request before request end', (done) => {
-      const requestUrl = '/requestUrl'
-      let requestReceivedByServer = false
-      server.on('request', (request, response) => {
-        switch (request.url) {
-          case requestUrl:
-            requestReceivedByServer = true
-            cancelRequest()
-            break
-          default:
-            handleUnexpectedURL(request, response)
-        }
-      })
-
-      let requestAbortEventEmitted = false
-      let requestCloseEventEmitted = false
-
-      const urlRequest = net.request({
-        method: 'GET',
-        url: `${server.url}${requestUrl}`
-      })
-      urlRequest.on('response', (response) => {
-        expect.fail('Unexpected response event')
-      })
-      urlRequest.on('finish', () => {
-        expect.fail('Unexpected finish event')
-      })
-      urlRequest.on('error', () => {
-        expect.fail('Unexpected error event')
-      })
-      urlRequest.on('abort', () => {
-        requestAbortEventEmitted = true
-      })
-      urlRequest.on('close', () => {
-        requestCloseEventEmitted = true
-        expect(requestReceivedByServer).to.be.true()
-        expect(requestAbortEventEmitted).to.be.true()
-        expect(requestCloseEventEmitted).to.be.true()
-        done()
-      })
-
-      urlRequest.chunkedEncoding = true
-      urlRequest.write(randomString(kOneKiloByte))
-      function cancelRequest () {
-        urlRequest.abort()
-      }
-    })
-
     it('it should be able to abort an HTTP request after request end and before response', (done) => {
       const requestUrl = '/requestUrl'
       let requestReceivedByServer = false
